Technical/Add FactoryBot linter, Bullet, rspec config Added FactoryBot linter rake task Added Bullet n+1 linter Refactored RSpec config namespaces
