+++ spec/support/config/json_matchers.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+JsonMatchers.schema_root = 'spec/support/schemas'
