+++ spec/support/matchers/not_change.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+RSpec::Matchers.define_negated_matcher :not_change, :change
